Add regex and fixed return

+++ loggin_in.py
@@ -1,21 +1,24 @@
 import hashlib
 import time
-
+import re
 def sign_up():
     email = input("Please input your e-mail: ")
     password = input ("Please input your password: ")
     confirm_password = input("Input your password again: ")
 
-    if password == confirm_password:
+    if password == confirm_password and re.match(".*@.*\..*", email):
         enc = confirm_password.encode()
         hash1 = hashlib.md5(enc).hexdigest()
         with open("logins.txt","w") as f:
             f.write(email + "\n")
             f.write(hash1)
         print("You have registered successfully! ")
-    #elif with wrong email
+    elif not re.match(".*@.*\..*", email):
+        print("You sure its an email?")
+        return
     else:
         print("Your passwords are not matching!")
+        return
 
     def your_data():
         print("Please input your data: ")
@@ -32,15 +35,15 @@ def sign_up():
         print("Alright then, hold up a second")
         time.sleep(5)
         your_data()
-    elif update in ["N"]:
+    elif update.upper() in ["N"]:
         pass
     else: 
         print("Something is wrong with you I can feel it")
 
     cont = input("Do you want to log in now?[Y/N]")
-    if cont in ["Y"]:
+    if cont.upper() in ["Y"]:
         log_in()
-    elif cont in ["N"]:
+    elif cont.upper() in ["N"]:
         print("Understandable, have a great day")
     else:
         print("You have no idea whats happening around you do you")
